backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Session==0.5.0
python-dotenv==1.0.0
requests==2.28.2
authlib==1.2.1